// backend/frontend/packages/promtube-backend/models/GetDanmaku200ResponseInner.ts
/**
 * Promtube Backend API
 * The API for Promtube
 *
 * OpenAPI spec version: 0.0.1
 * Contact: horahora1567@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class GetDanmaku200ResponseInner {
    'ID'?: number;
    'timestamp'?: string;
    'message'?: string;
    'authorID'?: number;
    'type'?: string;
    'color'?: string;
    'creationDate'?: string;
    'fontSize'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "ID",
            "baseName": "ID",
            "type": "number",
            "format": ""
        },
        {
            "name": "timestamp",
            "baseName": "Timestamp",
            "type": "string",
            "format": ""
        },
        {
            "name": "message",
            "baseName": "Message",
            "type": "string",
            "format": ""
        },
        {
            "name": "authorID",
            "baseName": "AuthorID",
            "type": "number",
            "format": ""
        },
        {
            "name": "type",
            "baseName": "Type",
            "type": "string",
            "format": ""
        },
        {
            "name": "color",
            "baseName": "Color",
            "type": "string",
            "format": ""
        },
        {
            "name": "creationDate",
            "baseName": "CreationDate",
            "type": "string",
            "format": ""
        },
        {
            "name": "fontSize",
            "baseName": "FontSize",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetDanmaku200ResponseInner.attributeTypeMap;
    }

    public constructor() {
    }
}